Code drop of security group rotation email bot
This implements a bot that emails the security group when new draft
advisories show up in the llvm security group repo. This bot @s people
who are currently oncall.

To this end, it also introduces a yaml file (and supporting Python
script) to define and extend the rotation.

For running this, Github Actions presents a few challenges:

1. All bot runs are public - observable changes in logs/etc could
   disclose security issues prior to us publishing them.
2. This requires non-committed state (mostly "what advisories have been
   emailed about already?")

So for now, the plan is just to run on one of my machines - I already
run llvmbb-monitor with reasonable uptime; adding to that isn't hard.

See https://github.com/gburgessiv/test-gha for development history
(though it's entirely just me hacking on my own with no input ;) )